requirements.txt
streamlit
pandas
matplotlib 
numpy 
pypdf
python-dotenv
groq
langchain
langchain-groq
langchain-huggingface
langchain-community
langchain-experimental 
langchain_cohere
langchain-chroma
transformers 
sentence_transformers
torch
chromadb
